requirements.txt
asgiref==3.8.1
click==8.1.7
Django==3.2.25
django-filebrowser==3.14.3
django-filter==23.5
django-grappelli==2.15.7
django-redis==5.4.0
django-rq==2.10.2
djangorestframework==3.15.1
drf-extensions==0.7.1
factory-boy==3.3.0
Faker==26.0.0
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pytest==8.3.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
rq==1.16.2
six==1.16.0
sqlparse==0.5.1
